Fix command line
