+++ pysqa/executor/__main__.py
@@ -1,58 +1,6 @@
-import os
 import sys
-
-from pympipool import PoolExecutor
-from pysqa.executor.helper import (
-    read_from_file,
-    deserialize,
-    apply_funct,
-    write_to_file,
-    serialize_result,
-)
-
-
-def execute_tasks(cores, cache_directory):
-    tasks_in_progress_dict = {}
-    with PoolExecutor(
-        max_workers=cores,
-        oversubscribe=False,
-        enable_flux_backend=False,
-        enable_slurm_backend=False,
-        init_function=None,
-        cwd=cache_directory,
-        queue_adapter=None,
-        queue_adapter_kwargs=None,
-    ) as exe:
-        while True:
-            file_lst = os.listdir(cache_directory)
-            for file_name_in in file_lst:
-                key = file_name_in.split(".in.pl")[0]
-                file_name_out = key + ".out.pl"
-                if (
-                    file_name_in.endswith(".in.pl")
-                    and file_name_out not in file_lst
-                    and file_name_out not in tasks_in_progress_dict.keys()
-                ):
-                    funct_dict = read_from_file(
-                        file_name=os.path.join(cache_directory, file_name_in)
-                    )
-                    apply_dict = deserialize(funct_dict=funct_dict)
-                    for k, v in apply_dict.items():
-                        tasks_in_progress_dict[k] = exe.submit(
-                            fn=v["fn"], args=v["args"], kwargs=v["kwargs"]
-                        )
-            for k, v in tasks_in_progress_dict.items():
-                if v.done():
-                    result_dict = apply_funct(apply_dict=v.result())
-                    write_to_file(
-                        funct_dict=serialize_result(result_dict=result_dict),
-                        state="out",
-                        cache_directory=cache_directory,
-                    )
+from pysqa.executor.backend import command_line
 
 
 if __name__ == "__main__":
-    arguments_lst = sys.argv
-    cores_arg = arguments_lst[arguments_lst.index("--cores") + 1]
-    path_arg = arguments_lst[arguments_lst.index("--path") + 1]
-    execute_tasks(cores=cores_arg, cache_directory=path_arg)
+    command_line(arguments_lst=sys.argv[1:])
